Add Vim Eunuch, fzf and update debug
Update vim debug settings to be more in-line with PHPStorm (and set the
IDE key as well).

Add Vim Eunuch, which adds nice file-based utilities to vim that one
might expect (including `:Delete`, which deletes a file from the file
system AND the buffer, avoiding annoying issues when creating a file
with the same name as a deleted one).

Add FZF.vim, and add the `Ctrl-O` mapping for fuzzy search within files.